# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: YNAB first-slice API
  version: 1.0.0
  description: >-
    The bounded slice supports cash and checking accounts, account lifecycle,
    same-budget transfers, categories, realized income, explicit income release,
    categorized spending, assignments, unassignments, moves, monthly summaries,
    dashboard values, and transaction history. Splits, cards, reconciliation, targets,
    scheduled transactions, future income, refunds, reimbursements, returns, and
    cleared/pending/uncleared states remain unsupported and are not part of this contract.
servers:
  - url: /
paths:
  /api/v1/auth/register:
    parameters: [{ $ref: '#/components/parameters/RequestId' }]
    post:
      summary: Register
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/RegisterInput' } } } }
      responses:
        '201': { $ref: '#/components/responses/UserSuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/auth/sign-in:
    parameters: [{ $ref: '#/components/parameters/RequestId' }]
    post:
      summary: Sign in
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/SignInInput' } } } }
      responses:
        '200': { $ref: '#/components/responses/SessionSuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
  /api/v1/auth/sign-out:
    parameters: [{ $ref: '#/components/parameters/RequestId' }]
    post:
      summary: Revoke the session cookie when present
      description: The sid cookie is optional; the server returns success and clears it.
      responses:
        '200': { $ref: '#/components/responses/EmptySuccess' }
  /api/v1/budgets:
    parameters: [{ $ref: '#/components/parameters/RequestId' }]
    get:
      summary: Resume the authenticated owner's budget
      security: [{ cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/BudgetSuccess' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      summary: Create the authenticated owner's budget
      security: [{ cookieAuth: [] }]
      responses:
        '201': { $ref: '#/components/responses/BudgetSuccess' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
      description: HttpOnly SameSite=Lax Path=/ session cookie; Secure is added in production.
  parameters:
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      description: Optional caller correlation ID; every success and error envelope returns requestId.
      schema: { type: string }
...
  responses:
    UserSuccess: { description: Registered user, content: { application/json: { schema: { $ref: '#/components/schemas/UserEnvelope' } } } }
    SessionSuccess: { description: Authenticated session; also sets the sid cookie, headers: { Set-Cookie: { description: Session cookie, schema: { type: string } } }, content: { application/json: { schema: { $ref: '#/components/schemas/SessionEnvelope' } } } }
    EmptySuccess: { description: Session revoked and cookie cleared, headers: { Set-Cookie: { description: Cleared session cookie, schema: { type: string } } }, content: { application/json: { schema: { $ref: '#/components/schemas/EmptyEnvelope' } } } }
    BudgetSuccess: { description: Budget state, content: { application/json: { schema: { $ref: '#/components/schemas/BudgetEnvelope' } } } }
...
    ValidationError: { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Unauthenticated: { description: Authentication required or credentials invalid, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Conflict: { description: State or concurrency conflict, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [data, requestId]
      properties:
        data: {}
        requestId: { type: string, minLength: 1, description: Non-empty request ID }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string, enum: [UNAUTHENTICATED, NOT_FOUND, CONFLICT, VALIDATION_ERROR, INTERNAL_ERROR] }
            message: { type: string }
            requestId: { type: string, minLength: 1, description: Non-empty request ID }
    User: { type: object, required: [id, email], properties: { id: { type: string }, email: { type: string, format: email } } }
    Session: { type: object, required: [user, sessionToken], properties: { user: { $ref: '#/components/schemas/User' }, sessionToken: { type: string, description: Opaque session token returned for this first slice } } }
    Account: { type: object, required: [id, name, kind, archived, openingBalanceMinor, balanceMinor], properties: { id: { type: string }, name: { type: string }, kind: { type: string, enum: [CASH, CHECKING] }, archived: { type: boolean }, openingBalanceMinor: { type: integer }, balanceMinor: { type: integer } } }
...
    Category: { type: object, required: [id, name, archived], properties: { id: { type: string }, name: { type: string }, archived: { type: boolean } } }
    Budget: { type: object, required: [id, setupStep, timezone, version, accounts, accountBalanceMinor, account, categories], properties: { id: { type: string }, setupStep: { type: string, enum: [ACCOUNT, CATEGORIES, COMPLETE] }, timezone: { type: string, enum: [UTC] }, version: { type: integer }, accounts: { type: array, items: { $ref: '#/components/schemas/Account' } }, accountBalanceMinor: { type: integer }, account: { nullable: true, allOf: [{ $ref: '#/components/schemas/Account' }] }, categories: { type: array, items: { $ref: '#/components/schemas/Category' } } } }
...
    RegisterInput: { type: object, required: [email, password], properties: { email: { type: string, format: email }, password: { type: string, minLength: 8 } } }
    SignInInput: { $ref: '#/components/schemas/RegisterInput' }
...
    UserEnvelope: { allOf: [{ $ref: '#/components/schemas/SuccessEnvelope' }], properties: { data: { $ref: '#/components/schemas/User' } } }
    SessionEnvelope: { allOf: [{ $ref: '#/components/schemas/SuccessEnvelope' }], properties: { data: { $ref: '#/components/schemas/Session' } } }
    EmptyEnvelope: { allOf: [{ $ref: '#/components/schemas/SuccessEnvelope' }], properties: { data: { nullable: true } } }
    BudgetEnvelope: { allOf: [{ $ref: '#/components/schemas/SuccessEnvelope' }], properties: { data: { $ref: '#/components/schemas/Budget' } } }
...